+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+- Add Azure RangeReaders [#386](https://github.com/geotrellis/geotrellis-server/pull/386)
+
 ## [4.5.0] - 2021-07-10
 
 ## Added