+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 - Add Azure RangeReaders [#386](https://github.com/geotrellis/geotrellis-server/pull/386)
+- Add GetFeatureInfoExtended support [#388](https://github.com/geotrellis/geotrellis-server/pull/388)
 
 ## [4.5.0] - 2021-07-10